+++ libraries/resource_private_internet_access_app_mac_os_x.rb
@@ -1,9 +1,10 @@
-# Encoding: UTF-8
+# encoding: utf-8
+# frozen_string_literal: true
 #
 # Cookbook Name:: private-internet-access
 # Library:: resource_private_internet_access_app_mac_os_x
 #
-# Copyright 2014-2015 Jonathan Hartman
+# Copyright 2014-2017, Jonathan Hartman
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -46,7 +47,7 @@ class Chef
         execute 'Run PIA installer' do
           command ::File.join(Chef::Config[:file_cache_path],
                               'Private\\ Internet\\ Access\\ Installer.app',
-                              'Contents/MacOS/runner')
+                              'Contents/MacOS/runner.sh')
           creates PATH
         end
       end