+++ spec/recipes/default_spec.rb
@@ -1,13 +1,15 @@
-# Encoding: UTF-8
+# encoding: utf-8
+# frozen_string_literal: true
 
 require 'spec_helper'
 
 describe 'private-internet-access::default' do
+  let(:platform) { { platform: 'mac_os_x', version: '10.10' } }
   let(:source) { nil }
   let(:runner) do
-    ChefSpec::ServerRunner.new do |node|
+    ChefSpec::ServerRunner.new(platform) do |node|
       unless source.nil?
-        node.set['private_internet_access']['app']['source'] = source
+        node.normal['private_internet_access']['app']['source'] = source
       end
     end
   end