+++ spec/resources/private_internet_access/mac_os_x/10_10_spec.rb
@@ -1,56 +1,12 @@
-require_relative '../../../spec_helper'
+# encoding: utf-8
+# frozen_string_literal: true
 
-describe 'resource_private_internet_access::mac_os_x::10_10' do
-  let(:source) { nil }
-  let(:action) { nil }
-  let(:runner) do
-    ChefSpec::SoloRunner.new(
-      step_into: 'private_internet_access',
-      platform: 'mac_os_x',
-      version: '10.10'
-    ) do |node|
-      unless source.nil?
-        node.set['private_internet_access']['app']['source'] = source
-      end
-    end
-  end
-  let(:converge) { runner.converge("private_internet_access_test::#{action}") }
+require_relative '../mac_os_x'
 
-  context 'the default action (:create)' do
-    let(:action) { :default }
+describe 'resources::private_internet_access::mac_os_x::10_10' do
+  include_context 'resources::private_internet_access::mac_os_x'
 
-    shared_examples_for 'any attribute set' do
-      it 'creates PIA' do
-        expect(chef_run).to create_private_internet_access('default')
-      end
+  let(:platform_version) { '10.10' }
 
-      it 'installs the PIA app' do
-        expect(chef_run).to install_private_internet_access_app('default')
-          .with(source: source)
-      end
-    end
-
-    context 'no source attribute' do
-      let(:source) { nil }
-      cached(:chef_run) { converge }
-
-      it_behaves_like 'any attribute set'
-    end
-
-    context 'a source attribute' do
-      let(:source) { 'http://example.com/pia.dmg' }
-      cached(:chef_run) { converge }
-
-      it_behaves_like 'any attribute set'
-    end
-  end
-
-  context 'the :remove action' do
-    let(:action) { :remove }
-    cached(:chef_run) { converge }
-
-    it 'removes the PIA app' do
-      expect(chef_run).to remove_private_internet_access('default')
-    end
-  end
+  it_behaves_like 'any Mac OS X platform'
 end