+++ spec/spec_helper.rb
@@ -1,4 +1,5 @@
-# Encoding: UTF-8
+# encoding: utf-8
+# frozen_string_literal: true
 
 require 'chef'
 require 'chefspec'
@@ -11,13 +12,16 @@ require 'coveralls'
 
 RSpec.configure do |c|
   c.color = true
+  c.file_cache_path = '/tmp'
 end
 
-SimpleCov.formatter = SimpleCov::Formatter::MultiFormatter[
-  Coveralls::SimpleCov::Formatter,
-  SimpleCov::Formatter::HTMLFormatter,
-  SimpleCov::Formatter::Console
-]
+SimpleCov.formatter = SimpleCov::Formatter::MultiFormatter.new(
+  [
+    Coveralls::SimpleCov::Formatter,
+    SimpleCov::Formatter::HTMLFormatter,
+    SimpleCov::Formatter::Console
+  ]
+)
 SimpleCov.minimum_coverage(100)
 SimpleCov.start