+++ spec/support/cookbooks/private_internet_access_app_test/recipes/remove.rb
@@ -1,5 +1,0 @@
-# Encoding: UTF-8
-
-private_internet_access_app 'default' do
-  action :remove
-end